Create a custom extension
